🚚 [RENAME] #20 Storybook 샘플/에셋 이동 및 설정 경로 반영
