+++ .storybook/samples/Header.stories.ts
@@ -1,8 +1,7 @@
+import { Header } from './Header';
 import type { Meta, StoryObj } from '@storybook/nextjs-vite';
 
 import { fn } from 'storybook/test';
-
-import { Header } from '../Header';
 
 const meta = {
   title: 'Example/Header',
