+++ .storybook/samples/assets/avif-test-image.avif
