+++ .storybook/samples/assets/figma-plugin.png
